refactor: restructure the code
